Refactoring in preparation of TimePicker implementation.
The Wxpython time and date pickers behave very similarly. This
refactoring extracts most of the boilerplate code into a base class that
will be shared between the two widgets.
